se arreglo NavBar, se creo HomePage y se utilizo outlet
